Fix connecting user from background thread (#2762)
* Change the DemoApp to always connect the user from BG Thread

This will make sure we test this often and don't break it

* Remove lazy initialisation of ChatClient to avoid thread-safe issues

* Change DefaultRequestEncoder to reference type

* Add comment explaining why we call connect from bg thread in the Demo App

* Fix unit tests and make sure current user id is thread-safe

* Simplify excluding current user from typing events

* Update CHANGELOG.md

* Improve DemoApp testing scenario

* Make ChatClient dependencies as constants

* Cleanup ChatClient initialization

* Move ChatClientFactory and Environment to separate files

* Fix tests not compiling
