+++ DemoApp/Shared/StreamChatWrapper.swift
@@ -76,14 +76,19 @@ extension StreamChatWrapper {
         // Setup Stream Chat
         setUpChat()
 
-        // Connect to chat
-        connect(user: user) { [weak self] in
-            if let error = $0 {
-                log.warning(error.localizedDescription)
-            } else {
-                self?.onRemotePushRegistration?()
+        // We connect from a background thread to make sure it works without issues/crashes.
+        // This is for testing purposes only. As a customer you can connect directly without dispatching to any queue.
+        DispatchQueue.global().async {
+            self.connect(user: user) { [weak self] error in
+                if let error = error {
+                    log.warning(error.localizedDescription)
+                } else {
+                    self?.onRemotePushRegistration?()
+                }
+                DispatchQueue.main.async {
+                    completion(error)
+                }
             }
-            completion($0)
         }
     }
