+++ Tests/StreamChatTests/APIClient/ChatPushNotificationContent_Tests.swift
@@ -36,7 +36,7 @@ final class ChatPushNotificationContent_Tests: XCTestCase {
 
         var env = ChatClient.Environment()
         env.databaseContainerBuilder = { _, _, _, _, _, _ in self.database }
-        env.apiClientBuilder = { _, _, _, _, _, _ in self.apiClient }
+        env.apiClientBuilder = { _, _, _, _ in self.apiClient }
         env.extensionLifecycleBuilder = { _ in self.extensionLifecycle }
         env.messageRepositoryBuilder = { _, _ in self.messageRepository }
